Move decorators to standalone file
